+++ src/legoify/__about__.py
@@ -1,4 +1,4 @@
 # SPDX-FileCopyrightText: 2026-present russ3llc <russellcarpenella@gmail.com>
 #
 # SPDX-License-Identifier: MIT
-__version__ = "0.5.7"
+__version__ = "0.6.0"
